+++ jest.config.js
@@ -33,5 +33,12 @@ module.exports = {
       lines: 100,
       statements: 100
     }
-  }
+  },
+  reporters: [
+    'default',
+    ['jest-junit', {
+      outputDirectory: 'test-results/unit',
+      outputName: 'test-results.xml'
+    }]
+  ],
 };